+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 # Confluent's Python client for Apache Kafka
 
+## v2.2.0
+
+v2.2.0 is a feature release with the following features, fixes and enhancements:
+
+ - [KIP-554](https://cwiki.apache.org/confluence/display/KAFKA/KIP-554%3A+Add+Broker-side+SCRAM+Config+API):
+   User SASL/SCRAM credentials alteration and description (#1575).
+
+confluent-kafka-python is based on librdkafka v2.2.0, see the
+[librdkafka release notes](https://github.com/edenhill/librdkafka/releases/tag/v2.2.0)
+for a complete list of changes, enhancements, fixes and upgrade considerations.
+
+
 ## v2.1.1
 
 v2.1.1 is a maintenance release with the following fixes and enhancements:
-
 
 ### Fixes